Add pure Android LocationRepository
