+++ app/src/main/kotlin/io/trewartha/positional/di/viewmodel/RepositoryModule.kt
@@ -1,41 +1,32 @@
 package io.trewartha.positional.di.viewmodel
 
-import android.content.Context
-import android.hardware.SensorManager
-import com.google.android.gms.location.FusedLocationProviderClient
+import dagger.Binds
 import dagger.Module
-import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewModelComponent
-import dagger.hilt.android.qualifiers.ApplicationContext
 import io.trewartha.positional.domain.compass.AndroidCompassReadingsRepository
 import io.trewartha.positional.domain.compass.CompassReadingsRepository
-import io.trewartha.positional.domain.location.AndroidFusedLocationRepository
+import io.trewartha.positional.domain.location.GmsLocationRepository
 import io.trewartha.positional.domain.location.LocationRepository
-import io.trewartha.positional.domain.settings.DataStoreSettingsRepository
-import io.trewartha.positional.domain.settings.SettingsRepository
 import io.trewartha.positional.domain.solunar.LocalSolarTimesRepository
 import io.trewartha.positional.domain.solunar.SolarTimesRepository
-import kotlinx.coroutines.Dispatchers
-import javax.inject.Singleton
 
 @Module
 @InstallIn(ViewModelComponent::class)
-class RepositoryModule {
+interface RepositoryModule {
 
-    @Provides
+    @Binds
     fun compassReadingsRepository(
-        sensorManager: SensorManager
-    ): CompassReadingsRepository = AndroidCompassReadingsRepository(sensorManager)
+        androidCompassReadingsRepository: AndroidCompassReadingsRepository
+    ): CompassReadingsRepository
 
-    @Provides
+    @Binds
     fun locationRepository(
-        fusedLocationProviderClient: FusedLocationProviderClient
-    ): LocationRepository = AndroidFusedLocationRepository(
-        coroutineDispatcher = Dispatchers.IO,
-        fusedLocationProviderClient = fusedLocationProviderClient
-    )
+        gmsLocationRepository: GmsLocationRepository
+    ): LocationRepository
 
-    @Provides
-    fun solarTimesRepository(): SolarTimesRepository = LocalSolarTimesRepository()
+    @Binds
+    fun solarTimesRepository(
+        localSolarTimesRepository: LocalSolarTimesRepository
+    ): SolarTimesRepository
 }
